+++ src/platform/DBlmdb.cpp
@@ -40,8 +40,12 @@ platform::lmdb::Txn::Txn(Env &db_env) {
 }
 
 void platform::lmdb::Txn::commit() {
-	lmdb_check(::mdb_txn_commit(handle), "mdb_txn_commit ");
-	handle = nullptr;
+    try {
+        lmdb_check(::mdb_txn_commit(handle), "mdb_txn_commit ");
+        handle = nullptr;
+    } catch (std::exception &e) {
+        std::cout << "commit erorr -> " << e.what() << std::endl;
+    }
 }
 
 platform::lmdb::Txn::~Txn() {